R package use a new version of survival Bcor

+++ src/surv.h
@@ -19,7 +19,7 @@
 
 
 void SRCT(double *x, double *t, int *delta, double *Sc, int *n, double *RCTV);
-void SRCT_new(int *x, int *t, int *delta, double *Sc, int *n, double *RCTV);
+void SRCT_new(double *x, int *t, int *delta, double *Sc, int *n, double *RCTV);
 
 
 #endif /* SURV_H_ */
